Добавляет исправления.
